support Webpack to Browserify
BrowserifyのほうがTypescriptがやりやすいので、browserifyにする。
