+++ index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="bower_components/balloon/balloon.css">
     <link rel="stylesheet" href="lib/fontawesome/css/font-awesome.css">
     <link rel="stylesheet" href="css/ui.css">
-    <script type="text/javascript" src="js/wpack.js"></script>
+    <script type="text/javascript" src="js/all.js"></script>
   </head>
   <body>
     <div class="loading"></div>
