+++ index.jade
@@ -7,7 +7,7 @@ html(lang="ja")
     link(rel="stylesheet", href="bower_components/balloon/balloon.css")
     link(rel="stylesheet", href="lib/fontawesome/css/font-awesome.css")
     link(rel="stylesheet", href="css/ui.css")
-    script(type='text/javascript', src='js/wpack.js')
+    script(type='text/javascript', src='js/all.js')
   body
     .loading
     canvas#pla-canvas(width='300', height='200')
