+++ package.json
@@ -4,11 +4,13 @@
   "description": "",
   "main": "n/a",
   "dependencies": {
+    "browserify": "^12.0.1",
     "glob": "^6.0.2",
     "gulp": "^3.9.0",
     "gulp-jade": "^1.1.0",
     "gulp-less": "^3.0.5",
     "gulp-typescript": "^2.10.0",
+    "tsify": "^0.13.1",
     "typescript": "^1.7.3"
   },
   "scripts": {
@@ -17,10 +19,11 @@
   "author": "shundroid",
   "license": "ISC",
   "devDependencies": {
-    "gulp-find-glob": "^1.0.1",
+    "del": "^2.2.0",
     "gulp-uglify": "^1.5.1",
-    "gulp-webpack": "^1.5.0",
-    "ts-loader": "^0.7.2",
-    "webpack": "^1.12.9"
+    "minimist": "^1.2.0",
+    "tsify": "^0.13.1",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.1.0"
   }
 }
